+++ example/spec/parallel/granuleInvalidator/CrossCollectionSpec.js
@@ -105,6 +105,7 @@ describe('The granule-invalidator deployed within a Cumulus workflow', () => {
         },
       ],
       cmrProvider: 'CUMULUS',
+      metricsProvider: 'CUMULUS',
     };
 
     await createCollection({ prefix: config.stackName,