+++ example/spec/parallel/testAPI/bulkDeleteSuccessSpec.js
@@ -301,6 +301,7 @@ describe('POST /granules/bulkDelete', () => {
         productionDateTime: message.record.productionDateTime,
         beginningDateTime: message.record.beginningDateTime,
         lastUpdateDateTime: message.record.lastUpdateDateTime,
+        metricsProvider: get(collection, 'metricsProvider'),
         cmrProvider: get(collection, 'cmrProvider'),
       };
       delete expectedGranuleAfterDeletion.cmrLink;