+++ packages/api/tests/endpoints/test-executions.js
@@ -183,6 +183,7 @@ test.before(async (t) => {
     name: collectionName,
     version: collectionVersion,
   });
+  t.context.metricsProvider = testPgCollection.metrics_provider;
   t.context.cmrProvider = testPgCollection.cmr_provider;
   const [pgCollection] = await t.context.collectionPgModel.create(
     knex,
@@ -1229,6 +1230,7 @@ test.serial('POST /executions publishes message to SNS topic', async (t) => {
 
   t.deepEqual(executionRecord, {
     ...translatedExecution,
+    metricsProvider: t.context.metricsProvider,
     cmrProvider: t.context.cmrProvider,
   });
 });
@@ -1579,6 +1581,7 @@ test.serial('PUT /executions publishes message to SNS topic', async (t) => {
   );
   t.deepEqual(executionRecord, {
     ...translatedExecution,
+    metricsProvider: t.context.metricsProvider,
     cmrProvider: t.context.cmrProvider,
     updatedAt: executionRecord.updatedAt,
   });