+++ packages/api/tests/lib/writeRecords/test-write-execution.js
@@ -608,6 +608,7 @@ test.serial('writeExecutionRecordFromMessage() successfully publishes an SNS mes
   t.is(executionRecord.status, cumulusMessage.meta.status);
   t.deepEqual(executionRecord, {
     ...translatedExecution,
+    metricsProvider: '',
     cmrProvider: '',
   });
 });
@@ -643,11 +644,12 @@ test.serial('writeExecutionRecordFromApi() successfully publishes an SNS message
   t.is(executionRecord.status, cumulusMessage.meta.status);
   t.deepEqual(executionRecord, {
     ...translatedExecution,
+    metricsProvider: '',
     cmrProvider: '',
   });
 });
 
-test.serial('writeExecutionRecordFromMessage() successfully publishes an SNS message with a collection and cmr provider', async (t) => {
+test.serial('writeExecutionRecordFromMessage() successfully publishes an SNS message with a collection and metricsProvider', async (t) => {
   const {
     cumulusMessage,
     executionArn,
@@ -655,7 +657,7 @@ test.serial('writeExecutionRecordFromMessage() successfully publishes an SNS mes
     knex,
     QueueUrl,
   } = t.context;
-  const collection = fakeCollectionRecordFactory({ cmr_provider: 'thisisacmrprovider' });
+  const collection = fakeCollectionRecordFactory({ metrics_provider: 'thisisamission', cmr_provider: 'thisisaprovider' });
   const collectionPgModel = new CollectionPgModel();
   const [collectionRecord] = await collectionPgModel.create(t.context.knex, collection);
   await writeExecutionRecordFromMessage({
@@ -680,6 +682,7 @@ test.serial('writeExecutionRecordFromMessage() successfully publishes an SNS mes
   t.is(executionRecord.status, cumulusMessage.meta.status);
   t.deepEqual(executionRecord, {
     ...translatedExecution,
+    metricsProvider: collection.metrics_provider,
     cmrProvider: collection.cmr_provider,
   });
 });