+++ packages/db/src/test-utils.ts
@@ -97,6 +97,7 @@ export const fakeCollectionRecordFactory = (
   created_at: new Date(),
   updated_at: new Date(),
   cmr_provider: 'provider',
+  metrics_provider: 'metrics_provider',
   ...params,
 });