+++ packages/db/tests/translate/test-collections.js
@@ -28,6 +28,7 @@ test('translatePostgresCollectionToApiCollection converts Postgres collection to
 
   const expected = {
     cmrProvider: 'provider',
+    metricsProvider: 'metrics_provider',
     createdAt: collectionRecord.created_at.getTime(),
     updatedAt: collectionRecord.updated_at.getTime(),
     files: JSON.parse(collectionRecord.files),